Split multi output stuff from single output
