+++ gpflow/multioutput/__init__.py
@@ -1,0 +1,25 @@
+# Copyright 2018 GPflow authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# flake8: noqa
+
+from .kernels import Mok
+from .kernels import SharedIndependentMok
+from .kernels import SeparateIndependentMok
+from .kernels import SeparateMixedMok
+
+from .features import Mof
+from .features import SharedIndependentMof
+from .features import SeparateIndependentMof
+from .features import MixedKernelSharedMof
