+++ gpflow/multioutput/kernels.py
@@ -1,15 +1,30 @@
+# Copyright 2018 GPflow authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
 import tensorflow as tf
 
-from . import settings
-from .decors import params_as_tensors, autoflow
-from .features import InducingPoints, InducingFeature
-from .features import Kuu, Kuf
-from .kernels import Kernel, Combination
-from . import kernels
-from .params import Parameter
-from .dispatch import dispatch
+from .. import kernels
+from .. import settings
+from ..decors import params_as_tensors, autoflow
+from ..dispatch import dispatch
+from ..features import InducingPoints, InducingFeature
+from ..features import Kuu, Kuf
+from ..kernels import Kernel, Combination
+from ..params import Parameter
 
-from .multifeatures import SeparateIndependentMof, SharedIndependentMof, MixedKernelSharedMof
+from .features import SeparateIndependentMof, SharedIndependentMof, MixedKernelSharedMof
 
 float_type = settings.float_type
